First take on ObjectHookLink
