--- Sources/InterposeKit/Internal/HookFinder.swift
@@ -1,59 +1,0 @@
-import Foundation
-
-extension Interpose {
-
-    private struct AssociatedKeys {
-        static var hookForBlock: UInt8 = 0
-    }
-
-    private class WeakObjectContainer<T: AnyObject>: NSObject {
-        private weak var _object: T?
-
-        var object: T? {
-            return _object
-        }
-        init(with object: T?) {
-            _object = object
-        }
-    }
-
-    static func storeHook(hook: Hook, to imp: IMP) {
-        // Weakly store reference to hook inside the block of the IMP.
-        guard let block = imp_getBlock(imp) else { fatalError() }
-        
-        objc_setAssociatedObject(
-            block,
-            &AssociatedKeys.hookForBlock,
-            WeakObjectContainer(with: hook),
-            .OBJC_ASSOCIATION_RETAIN
-        )
-
-    }
-
-    // Finds the hook to a given implementation.
-    static func hookForIMP(_ imp: IMP) -> Hook? {
-        // Get the block that backs our IMP replacement
-        guard let block = imp_getBlock(imp) else { return nil }
-        let container = objc_getAssociatedObject(block, &AssociatedKeys.hookForBlock) as? WeakObjectContainer<Hook>
-        return container?.object
-    }
-
-    // Find the hook above us (not necessarily topmost)
-    static func findNextHook(selfHook: Hook, topmostIMP: IMP) -> Hook? {
-        // We are not topmost hook, so find the hook above us!
-        var impl: IMP? = topmostIMP
-        var currentHook: Hook?
-        repeat {
-            // get topmost hook
-            let hook: Hook? = Interpose.hookForIMP(impl!)
-            if hook === selfHook {
-                // return parent
-                return currentHook
-            }
-            // crawl down the chain until we find ourselves
-            currentHook = hook
-            impl = hook?.strategy.originalIMP
-        } while impl != nil
-        return nil
-    }
-}
